🧪 move tests to separate dir
